calculo de precio total en pedidos

+++ BogoThrashManager/src/main/java/co/edu/poli/BogoThrashManager/RegistroPedidos/modelo/DetallePedido.java
@@ -53,6 +53,6 @@ public class DetallePedido {
      * 
      */
 	@Column(name = "precioTotal")
-    private Long precioTotal;
+    private Double precioTotal;
 
 }
